Crate google-identitytoolkit3

This documentation was generated from Identity Toolkit crate version 0.0.1. The original source code can be found on github.

Features

It seems there is nothing you can do here ... .

relyingparty

create auth uri, delete account, download account, get account info, get oob confirmation code, get public keys, reset password, set account info, upload account, verify assertion and verify password

Everything else about the Identity Toolkit v3 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.relyingparty().verify_assertion(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-identitytoolkit3 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-identitytoolkit3" as identitytoolkit3;
use identitytoolkit3::IdentitytoolkitRelyingpartyVerifyAssertionRequest;
use identitytoolkit3::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use identitytoolkit3::IdentityToolkit;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = IdentityToolkit::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: IdentitytoolkitRelyingpartyVerifyAssertionRequest = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.relyingparty().verify_assertion(&req)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

CreateAuthUriResponse
Response of creating the IDP authentication URL.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeleteAccountResponse
Respone of deleting account.

DownloadAccountResponse
Respone of downloading accounts in batch.

GetAccountInfoResponse
Response of getting account information.

GetOobConfirmationCodeResponse
Response of getting a code for user confirmation (reset password, change email etc.).

IdentityToolkit
Central instance to access all IdentityToolkit related resource activities

IdentitytoolkitRelyingpartyCreateAuthUriRequest
Request to get the IDP authentication URL.

IdentitytoolkitRelyingpartyDeleteAccountRequest
Request to delete account.

IdentitytoolkitRelyingpartyDownloadAccountRequest
Request to download user account in batch.

IdentitytoolkitRelyingpartyGetAccountInfoRequest
Request to get the account information.

IdentitytoolkitRelyingpartyGetPublicKeysResponse
Respone of getting public keys.

IdentitytoolkitRelyingpartyResetPasswordRequest
Request to reset the password.

IdentitytoolkitRelyingpartySetAccountInfoRequest
Request to set the account information.

IdentitytoolkitRelyingpartyUploadAccountRequest
Request to upload user account in batch.

IdentitytoolkitRelyingpartyVerifyAssertionRequest
Request to verify the IDP assertion.

IdentitytoolkitRelyingpartyVerifyPasswordRequest
Request to verify the password.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Relyingparty
Request of getting a code for user confirmation (reset password, change email etc.)

RelyingpartyCreateAuthUriCallBuilder
Creates the URI used by the IdP to authenticate the user.

RelyingpartyDeleteAccountCallBuilder
Delete user account.

RelyingpartyDownloadAccountCallBuilder
Batch download user accounts.

RelyingpartyGetAccountInfoCallBuilder
Returns the account info.

RelyingpartyGetOobConfirmationCodeCallBuilder
Get a code for user action confirmation.

RelyingpartyGetPublicKeyCallBuilder
Get token signing public key.

RelyingpartyMethodsBuilder
A builder providing access to all methods supported on relyingparty resources. It is not used directly, but through the IdentityToolkit hub.

RelyingpartyResetPasswordCallBuilder
Reset password for a user.

RelyingpartySetAccountInfoCallBuilder
Set account info for a user.

RelyingpartyUploadAccountCallBuilder
Batch upload existing user accounts.

RelyingpartyVerifyAssertionCallBuilder
Verifies the assertion returned by the IdP.

RelyingpartyVerifyPasswordCallBuilder
Verifies the user entered password.

ResetPasswordResponse
Response of resetting the password.

SetAccountInfoResponse
Respone of setting the account information.

SetAccountInfoResponseProviderUserInfo
The user's profiles at the associated IdPs.

UploadAccountResponse
Respone of uploading accounts in batch.

UploadAccountResponseError
The error encountered while processing the account info.

UserInfo
Template for an individual account info.

UserInfoProviderUserInfo
The IDP of the user.

VerifyAssertionResponse
Response of verifying the IDP assertion.

VerifyPasswordResponse
Request of verifying the password.

Enums

Result
A universal result type used as return for all action method results.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.