Module google_identitytoolkit3::api

Structs

CreateAuthUriResponse
Response of creating the IDP authentication URL.

DeleteAccountResponse
Respone of deleting account.

DownloadAccountResponse
Response of downloading accounts in batch.

EmailLinkSigninResponse
Response of email signIn.

EmailTemplate
Template for an email template.

GetAccountInfoResponse
Response of getting account information.

GetOobConfirmationCodeResponse
Response of getting a code for user confirmation (reset password, change email etc.).

GetRecaptchaParamResponse
Response of getting recaptcha param.

IdentityToolkit
Central instance to access all IdentityToolkit related resource activities

IdentitytoolkitRelyingpartyCreateAuthUriRequest
Request to get the IDP authentication URL.

IdentitytoolkitRelyingpartyDeleteAccountRequest
Request to delete account.

IdentitytoolkitRelyingpartyDownloadAccountRequest
Request to download user account in batch.

IdentitytoolkitRelyingpartyEmailLinkSigninRequest
Request to sign in with email.

IdentitytoolkitRelyingpartyGetAccountInfoRequest
Request to get the account information.

IdentitytoolkitRelyingpartyGetProjectConfigResponse
Response of getting the project configuration.

IdentitytoolkitRelyingpartyGetPublicKeysResponse
Respone of getting public keys.

IdentitytoolkitRelyingpartyResetPasswordRequest
Request to reset the password.

IdentitytoolkitRelyingpartySendVerificationCodeRequest
Request for Identitytoolkit-SendVerificationCode

IdentitytoolkitRelyingpartySendVerificationCodeResponse
Response for Identitytoolkit-SendVerificationCode

IdentitytoolkitRelyingpartySetAccountInfoRequest
Request to set the account information.

IdentitytoolkitRelyingpartySetProjectConfigRequest
Request to set the project configuration.

IdentitytoolkitRelyingpartySetProjectConfigResponse
Response of setting the project configuration.

IdentitytoolkitRelyingpartySignOutUserRequest
Request to sign out user.

IdentitytoolkitRelyingpartySignOutUserResponse
Response of signing out user.

IdentitytoolkitRelyingpartySignupNewUserRequest
Request to signup new user, create anonymous user or anonymous user reauth.

IdentitytoolkitRelyingpartyUploadAccountRequest
Request to upload user account in batch.

IdentitytoolkitRelyingpartyVerifyAssertionRequest
Request to verify the IDP assertion.

IdentitytoolkitRelyingpartyVerifyCustomTokenRequest
Request to verify a custom token

IdentitytoolkitRelyingpartyVerifyPasswordRequest
Request to verify the password.

IdentitytoolkitRelyingpartyVerifyPhoneNumberRequest
Request for Identitytoolkit-VerifyPhoneNumber

IdentitytoolkitRelyingpartyVerifyPhoneNumberResponse
Response for Identitytoolkit-VerifyPhoneNumber

IdpConfig
Template for a single idp configuration.

Relyingparty
Request of getting a code for user confirmation (reset password, change email etc.)

RelyingpartyCreateAuthUriCall
Creates the URI used by the IdP to authenticate the user.

RelyingpartyDeleteAccountCall
Delete user account.

RelyingpartyDownloadAccountCall
Batch download user accounts.

RelyingpartyEmailLinkSigninCall
Reset password for a user.

RelyingpartyGetAccountInfoCall
Returns the account info.

RelyingpartyGetOobConfirmationCodeCall
Get a code for user action confirmation.

RelyingpartyGetProjectConfigCall
Get project configuration.

RelyingpartyGetPublicKeyCall
Get token signing public key.

RelyingpartyGetRecaptchaParamCall
Get recaptcha secure param.

RelyingpartyMethods
A builder providing access to all methods supported on relyingparty resources. It is not used directly, but through the IdentityToolkit hub.

RelyingpartyResetPasswordCall
Reset password for a user.

RelyingpartySendVerificationCodeCall
Send SMS verification code.

RelyingpartySetAccountInfoCall
Set account info for a user.

RelyingpartySetProjectConfigCall
Set project configuration.

RelyingpartySignOutUserCall
Sign out user.

RelyingpartySignupNewUserCall
Signup new user.

RelyingpartyUploadAccountCall
Batch upload existing user accounts.

RelyingpartyVerifyAssertionCall
Verifies the assertion returned by the IdP.

RelyingpartyVerifyCustomTokenCall
Verifies the developer asserted ID token.

RelyingpartyVerifyPasswordCall
Verifies the user entered password.

RelyingpartyVerifyPhoneNumberCall
Verifies ownership of a phone number and creates/updates the user account accordingly.

ResetPasswordResponse
Response of resetting the password.

SetAccountInfoResponse
Respone of setting the account information.

SetAccountInfoResponseProviderUserInfo
The user's profiles at the associated IdPs.

SignupNewUserResponse
Response of signing up new user, creating anonymous user or anonymous user reauth.

UploadAccountResponse
Respone of uploading accounts in batch.

UploadAccountResponseError
The error encountered while processing the account info.

UserInfo
Template for an individual account info.

UserInfoProviderUserInfo
The IDP of the user.

VerifyAssertionResponse
Response of verifying the IDP assertion.

VerifyCustomTokenResponse
Response from verifying a custom token

VerifyPasswordResponse
Request of verifying the password.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.