Struct google_identitytoolkit3::IdentitytoolkitRelyingpartyCreateAuthUriRequest

pub struct IdentitytoolkitRelyingpartyCreateAuthUriRequest {
    pub identifier: Option<String>,
    pub oauth_scope: Option<String>,
    pub client_id: Option<String>,
    pub ota_app: Option<String>,
    pub oauth_consumer_key: Option<String>,
    pub provider_id: Option<String>,
    pub context: Option<String>,
    pub app_id: Option<String>,
    pub continue_uri: Option<String>,
    pub openid_realm: Option<String>,
}

Request to get the IDP authentication URL.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

create auth uri relyingparty (request)

Fields

identifier

The email or federated ID of the user.

oauth_scope

Additional oauth scopes, beyond the basid user profile, that the user would be prompted to grant

client_id

The relying party OAuth client ID.

ota_app

The native app package for OTA installation.

oauth_consumer_key

The developer's consumer key for OpenId OAuth Extension

provider_id

The IdP ID. For white listed IdPs it's a short domain name e.g. google.com, aol.com, live.net and yahoo.com. For other OpenID IdPs it's the OP identifier.

context

The opaque value used by the client to maintain context info between the authentication request and the IDP callback.

app_id

The app ID of the mobile app, base64(CERT_SHA1):PACKAGE_NAME for Android, BUNDLE_ID for iOS.

continue_uri

The URI to which the IDP redirects the user after the federated login flow.

openid_realm

Optional realm for OpenID protocol. The sub string "scheme://domain:port" of the param "continueUri" is used if this is not set.