Struct google_identitytoolkit3::IdentitytoolkitRelyingpartySetAccountInfoRequest

pub struct IdentitytoolkitRelyingpartySetAccountInfoRequest {
    pub delegated_project_number: Option<String>,
    pub oob_code: Option<String>,
    pub local_id: Option<String>,
    pub valid_since: Option<String>,
    pub photo_url: Option<String>,
    pub instance_id: Option<String>,
    pub disable_user: Option<bool>,
    pub return_secure_token: Option<bool>,
    pub password: Option<String>,
    pub display_name: Option<String>,
    pub delete_attribute: Option<Vec<String>>,
    pub upgrade_to_federated_login: Option<bool>,
    pub id_token: Option<String>,
    pub email_verified: Option<bool>,
    pub delete_provider: Option<Vec<String>>,
    pub captcha_challenge: Option<String>,
    pub captcha_response: Option<String>,
    pub provider: Option<Vec<String>>,
    pub email: Option<String>,
}

Request to set the account information.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set account info relyingparty (request)

Fields

delegated_project_number

GCP project number of the requesting delegated app. Currently only intended for Firebase V1 migration.

oob_code

The out-of-band code of the change email request.

local_id

The local ID of the user.

valid_since

Timestamp in seconds for valid login token.

photo_url

The photo url of the user.

instance_id

Instance id token of the app.

disable_user

Whether to disable the user.

return_secure_token

Whether return sts id token and refresh token instead of gitkit token.

password

The new password of the user.

display_name

The name of the user.

delete_attribute

The attributes users request to delete.

upgrade_to_federated_login

Mark the user to upgrade to federated login.

id_token

The GITKit token of the authenticated user.

email_verified

Mark the email as verified or not.

delete_provider

The IDPs the user request to delete.

captcha_challenge

The captcha challenge.

captcha_response

Response to the captcha.

provider

The associated IDPs of the user.

email

The email of the user.