Struct google_identitytoolkit3::IdentitytoolkitRelyingpartyVerifyAssertionRequest

pub struct IdentitytoolkitRelyingpartyVerifyAssertionRequest {
    pub post_body: Option<String>,
    pub delegated_project_number: Option<String>,
    pub instance_id: Option<String>,
    pub id_token: Option<String>,
    pub pending_id_token: Option<String>,
    pub session_id: Option<String>,
    pub request_uri: Option<String>,
    pub return_secure_token: Option<bool>,
    pub return_refresh_token: Option<bool>,
}

Request to verify the IDP assertion.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

verify assertion relyingparty (request)

Fields

post_body

The post body if the request is a HTTP POST.

delegated_project_number

GCP project number of the requesting delegated app. Currently only intended for Firebase V1 migration.

instance_id

Instance id token of the app.

id_token

The GITKit token of the authenticated user.

pending_id_token

The GITKit token for the non-trusted IDP pending to be confirmed by the user.

session_id

Session ID, which should match the one in previous createAuthUri request.

request_uri

The URI to which the IDP redirects the user back. It may contain federated login result params added by the IDP.

return_secure_token

Whether return sts id token and refresh token instead of gitkit token.

return_refresh_token

Whether to return refresh tokens.