Struct google_identitytoolkit3::RelyingpartyMethods

pub struct RelyingpartyMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on relyingparty resources. It is not used directly, but through the IdentityToolkit hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_identitytoolkit3 as identitytoolkit3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use identitytoolkit3::IdentityToolkit;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = IdentityToolkit::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create_auth_uri(...)`, `delete_account(...)`, `download_account(...)`, `email_link_signin(...)`, `get_account_info(...)`, `get_oob_confirmation_code(...)`, `get_project_config(...)`, `get_public_keys(...)`, `get_recaptcha_param(...)`, `reset_password(...)`, `send_verification_code(...)`, `set_account_info(...)`, `set_project_config(...)`, `sign_out_user(...)`, `signup_new_user(...)`, `upload_account(...)`, `verify_assertion(...)`, `verify_custom_token(...)`, `verify_password(...)` and `verify_phone_number(...)`
// to build up your call.
let rb = hub.relyingparty();

Methods

impl<'a, C, A> RelyingpartyMethods<'a, C, A>

pub fn email_link_signin(
&self,
request: IdentitytoolkitRelyingpartyEmailLinkSigninRequest
) -> RelyingpartyEmailLinkSigninCall<'a, C, A>

Create a builder to help you perform the following task:

Reset password for a user.

Arguments

request - No description provided.

pub fn get_oob_confirmation_code(
&self,
request: Relyingparty
) -> RelyingpartyGetOobConfirmationCodeCall<'a, C, A>

Create a builder to help you perform the following task:

Get a code for user action confirmation.

Arguments

request - No description provided.

pub fn get_public_keys(&self) -> RelyingpartyGetPublicKeyCall<'a, C, A>

Create a builder to help you perform the following task:

Get token signing public key.

pub fn create_auth_uri(
&self,
request: IdentitytoolkitRelyingpartyCreateAuthUriRequest
) -> RelyingpartyCreateAuthUriCall<'a, C, A>

Create a builder to help you perform the following task:

Creates the URI used by the IdP to authenticate the user.

Arguments

request - No description provided.

pub fn get_recaptcha_param(&self) -> RelyingpartyGetRecaptchaParamCall<'a, C, A>

Create a builder to help you perform the following task:

Get recaptcha secure param.

pub fn sign_out_user(
&self,
request: IdentitytoolkitRelyingpartySignOutUserRequest
) -> RelyingpartySignOutUserCall<'a, C, A>

Create a builder to help you perform the following task:

Sign out user.

Arguments

request - No description provided.

pub fn verify_assertion(
&self,
request: IdentitytoolkitRelyingpartyVerifyAssertionRequest
) -> RelyingpartyVerifyAssertionCall<'a, C, A>

Create a builder to help you perform the following task:

Verifies the assertion returned by the IdP.

Arguments

request - No description provided.

pub fn upload_account(
&self,
request: IdentitytoolkitRelyingpartyUploadAccountRequest
) -> RelyingpartyUploadAccountCall<'a, C, A>

Create a builder to help you perform the following task:

Batch upload existing user accounts.

Arguments

request - No description provided.

pub fn get_account_info(
&self,
request: IdentitytoolkitRelyingpartyGetAccountInfoRequest
) -> RelyingpartyGetAccountInfoCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the account info.

Arguments

request - No description provided.

pub fn verify_custom_token(
&self,
request: IdentitytoolkitRelyingpartyVerifyCustomTokenRequest
) -> RelyingpartyVerifyCustomTokenCall<'a, C, A>

Create a builder to help you perform the following task:

Verifies the developer asserted ID token.

Arguments

request - No description provided.

pub fn reset_password(
&self,
request: IdentitytoolkitRelyingpartyResetPasswordRequest
) -> RelyingpartyResetPasswordCall<'a, C, A>

Create a builder to help you perform the following task:

Reset password for a user.

Arguments

request - No description provided.

pub fn download_account(
&self,
request: IdentitytoolkitRelyingpartyDownloadAccountRequest
) -> RelyingpartyDownloadAccountCall<'a, C, A>

Create a builder to help you perform the following task:

Batch download user accounts.

Arguments

request - No description provided.

pub fn set_account_info(
&self,
request: IdentitytoolkitRelyingpartySetAccountInfoRequest
) -> RelyingpartySetAccountInfoCall<'a, C, A>

Create a builder to help you perform the following task:

Set account info for a user.

Arguments

request - No description provided.

pub fn delete_account(
&self,
request: IdentitytoolkitRelyingpartyDeleteAccountRequest
) -> RelyingpartyDeleteAccountCall<'a, C, A>

Create a builder to help you perform the following task:

Delete user account.

Arguments

request - No description provided.

pub fn signup_new_user(
&self,
request: IdentitytoolkitRelyingpartySignupNewUserRequest
) -> RelyingpartySignupNewUserCall<'a, C, A>

Create a builder to help you perform the following task:

Signup new user.

Arguments

request - No description provided.

pub fn verify_phone_number(
&self,
request: IdentitytoolkitRelyingpartyVerifyPhoneNumberRequest
) -> RelyingpartyVerifyPhoneNumberCall<'a, C, A>

Create a builder to help you perform the following task:

Verifies ownership of a phone number and creates/updates the user account accordingly.

Arguments

request - No description provided.

pub fn verify_password(
&self,
request: IdentitytoolkitRelyingpartyVerifyPasswordRequest
) -> RelyingpartyVerifyPasswordCall<'a, C, A>

Create a builder to help you perform the following task:

Verifies the user entered password.

Arguments

request - No description provided.

pub fn set_project_config(
&self,
request: IdentitytoolkitRelyingpartySetProjectConfigRequest
) -> RelyingpartySetProjectConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Set project configuration.

Arguments

request - No description provided.

pub fn send_verification_code(
&self,
request: IdentitytoolkitRelyingpartySendVerificationCodeRequest
) -> RelyingpartySendVerificationCodeCall<'a, C, A>

Create a builder to help you perform the following task:

Send SMS verification code.

Arguments

request - No description provided.

pub fn get_project_config(&self) -> RelyingpartyGetProjectConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Get project configuration.