Struct google_identitytoolkit3::SetAccountInfoResponse

pub struct SetAccountInfoResponse {
    pub expires_in: Option<String>,
    pub kind: Option<String>,
    pub display_name: Option<String>,
    pub local_id: Option<String>,
    pub photo_url: Option<String>,
    pub id_token: Option<String>,
    pub password_hash: Option<String>,
    pub new_email: Option<String>,
    pub refresh_token: Option<String>,
    pub provider_user_info: Option<Vec<SetAccountInfoResponseProviderUserInfo>>,
    pub email: Option<String>,
}

Respone of setting the account information.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set account info relyingparty (response)

Fields
expires_in: Option<String>

If idToken is STS id token, then this field will be expiration time of STS id token in seconds.

kind: Option<String>

The fixed string "identitytoolkit#SetAccountInfoResponse".

display_name: Option<String>

The name of the user.

local_id: Option<String>

The local ID of the user.

photo_url: Option<String>

The photo url of the user.

id_token: Option<String>

The Gitkit id token to login the newly sign up user.

password_hash: Option<String>

The user's hashed password.

new_email: Option<String>

The new email the user attempts to change to.

refresh_token: Option<String>

If idToken is STS id token, then this field will be refresh token.

provider_user_info: Option<Vec<SetAccountInfoResponseProviderUserInfo>>

The user's profiles at the associated IdPs.

email: Option<String>

The email of the user.