Struct google_identitytoolkit3::IdentitytoolkitRelyingpartyCreateAuthUriRequest

pub struct IdentitytoolkitRelyingpartyCreateAuthUriRequest {
    pub hosted_domain: Option<String>,
    pub session_id: Option<String>,
    pub identifier: Option<String>,
    pub auth_flow_type: Option<String>,
    pub oauth_scope: Option<String>,
    pub client_id: Option<String>,
    pub ota_app: Option<String>,
    pub oauth_consumer_key: Option<String>,
    pub provider_id: Option<String>,
    pub context: Option<String>,
    pub app_id: Option<String>,
    pub continue_uri: Option<String>,
    pub openid_realm: Option<String>,
}

Request to get the IDP authentication URL.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

create auth uri relyingparty (request)

Fields
hosted_domain: Option<String>

The hosted domain to restrict sign-in to accounts at that domain for Google Apps hosted accounts.

session_id: Option<String>

The session_id passed by client.

identifier: Option<String>

The email or federated ID of the user.

auth_flow_type: Option<String>

Explicitly specify the auth flow type. Currently only support "CODE_FLOW" type. The field is only used for Google provider.

oauth_scope: Option<String>

Additional oauth scopes, beyond the basid user profile, that the user would be prompted to grant

client_id: Option<String>

The relying party OAuth client ID.

ota_app: Option<String>

The native app package for OTA installation.

oauth_consumer_key: Option<String>

The developer's consumer key for OpenId OAuth Extension

provider_id: Option<String>

The IdP ID. For white listed IdPs it's a short domain name e.g. google.com, aol.com, live.net and yahoo.com. For other OpenID IdPs it's the OP identifier.

context: Option<String>

The opaque value used by the client to maintain context info between the authentication request and the IDP callback.

app_id: Option<String>

The app ID of the mobile app, base64(CERT_SHA1):PACKAGE_NAME for Android, BUNDLE_ID for iOS.

continue_uri: Option<String>

The URI to which the IDP redirects the user after the federated login flow.

openid_realm: Option<String>

Optional realm for OpenID protocol. The sub string "scheme://domain:port" of the param "continueUri" is used if this is not set.