Struct google_identitytoolkit3::IdentitytoolkitRelyingpartySetAccountInfoRequest

pub struct IdentitytoolkitRelyingpartySetAccountInfoRequest {
    pub delegated_project_number: Option<String>,
    pub oob_code: Option<String>,
    pub local_id: Option<String>,
    pub valid_since: Option<String>,
    pub photo_url: Option<String>,
    pub instance_id: Option<String>,
    pub last_login_at: Option<String>,
    pub disable_user: Option<bool>,
    pub return_secure_token: Option<bool>,
    pub password: Option<String>,
    pub display_name: Option<String>,
    pub created_at: Option<String>,
    pub delete_attribute: Option<Vec<String>>,
    pub upgrade_to_federated_login: Option<bool>,
    pub id_token: Option<String>,
    pub email_verified: Option<bool>,
    pub delete_provider: Option<Vec<String>>,
    pub captcha_challenge: Option<String>,
    pub captcha_response: Option<String>,
    pub provider: Option<Vec<String>>,
    pub email: Option<String>,
}

Request to set the account information.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set account info relyingparty (request)

Fields
delegated_project_number: Option<String>

GCP project number of the requesting delegated app. Currently only intended for Firebase V1 migration.

oob_code: Option<String>

The out-of-band code of the change email request.

local_id: Option<String>

The local ID of the user.

valid_since: Option<String>

Timestamp in seconds for valid login token.

photo_url: Option<String>

The photo url of the user.

instance_id: Option<String>

Instance id token of the app.

last_login_at: Option<String>

Last login timestamp.

disable_user: Option<bool>

Whether to disable the user.

return_secure_token: Option<bool>

Whether return sts id token and refresh token instead of gitkit token.

password: Option<String>

The new password of the user.

display_name: Option<String>

The name of the user.

created_at: Option<String>

The timestamp when the account is created.

delete_attribute: Option<Vec<String>>

The attributes users request to delete.

upgrade_to_federated_login: Option<bool>

Mark the user to upgrade to federated login.

id_token: Option<String>

The GITKit token of the authenticated user.

email_verified: Option<bool>

Mark the email as verified or not.

delete_provider: Option<Vec<String>>

The IDPs the user request to delete.

captcha_challenge: Option<String>

The captcha challenge.

captcha_response: Option<String>

Response to the captcha.

provider: Option<Vec<String>>

The associated IDPs of the user.

email: Option<String>

The email of the user.