Struct google_identitytoolkit3::VerifyAssertionResponse

pub struct VerifyAssertionResponse {
    pub federated_id: Option<String>,
    pub local_id: Option<String>,
    pub photo_url: Option<String>,
    pub input_email: Option<String>,
    pub oauth_scope: Option<String>,
    pub email_verified: Option<bool>,
    pub need_confirmation: Option<bool>,
    pub app_scheme: Option<String>,
    pub full_name: Option<String>,
    pub app_installation_url: Option<String>,
    pub kind: Option<String>,
    pub display_name: Option<String>,
    pub language: Option<String>,
    pub first_name: Option<String>,
    pub oauth_authorization_code: Option<String>,
    pub oauth_expire_in: Option<i32>,
    pub id_token: Option<String>,
    pub verified_provider: Option<Vec<String>>,
    pub oauth_access_token: Option<String>,
    pub oauth_request_token: Option<String>,
    pub date_of_birth: Option<String>,
    pub original_email: Option<String>,
    pub provider_id: Option<String>,
    pub context: Option<String>,
    pub last_name: Option<String>,
    pub action: Option<String>,
    pub time_zone: Option<String>,
    pub nick_name: Option<String>,
    pub email: Option<String>,
    pub email_recycled: Option<bool>,
}

Response of verifying the IDP assertion.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

verify assertion relyingparty (response)

Fields

federated_id

The unique ID identifies the IdP account.

local_id

The RP local ID if it's already been mapped to the IdP account identified by the federated ID.

photo_url

The URI of the public accessible profiel picture.

input_email

It's the identifier param in the createAuthUri request if the identifier is an email. It can be used to check whether the user input email is different from the asserted email.

oauth_scope

The scope for the OpenID OAuth extension.

email_verified

The value is true if the IDP is also the email provider. It means the user owns the email.

need_confirmation

Whether the assertion is from a non-trusted IDP and need account linking confirmation.

app_scheme

The custom scheme used by mobile app.

full_name

The full name of the user.

app_installation_url

URL for OTA app installation.

kind

The fixed string "identitytoolkit#VerifyAssertionResponse".

display_name

The display name of the user.

language

The language preference of the user.

first_name

The first name of the user.

oauth_authorization_code

The OAuth2 authorization code.

oauth_expire_in

The lifetime in seconds of the OAuth2 access token.

id_token

The ID token.

verified_provider

When action is 'map', contains the idps which can be used for confirmation.

oauth_access_token

The OAuth2 access token.

oauth_request_token

The user approved request token for the OpenID OAuth extension.

date_of_birth

The birth date of the IdP account.

original_email

The original email stored in the mapping storage. It's returned when the federated ID is associated to a different email.

provider_id

The IdP ID. For white listed IdPs it's a short domain name e.g. google.com, aol.com, live.net and yahoo.com. If the "providerId" param is set to OpenID OP identifer other than the whilte listed IdPs the OP identifier is returned. If the "identifier" param is federated ID in the createAuthUri request. The domain part of the federated ID is returned.

context

The opaque value used by the client to maintain context info between the authentication request and the IDP callback.

last_name

The last name of the user.

action

The action code.

time_zone

The timezone of the user.

nick_name

The nick name of the user.

email

The email returned by the IdP. NOTE: The federated login user may not own the email.

email_recycled

It's true if the email is recycled.