Struct google_identitytoolkit3::VerifyAssertionResponse

pub struct VerifyAssertionResponse {
    pub federated_id: Option<String>,
    pub local_id: Option<String>,
    pub photo_url: Option<String>,
    pub is_new_user: Option<bool>,
    pub input_email: Option<String>,
    pub raw_user_info: Option<String>,
    pub app_scheme: Option<String>,
    pub expires_in: Option<String>,
    pub verified_provider: Option<Vec<String>>,
    pub need_email: Option<bool>,
    pub date_of_birth: Option<String>,
    pub first_name: Option<String>,
    pub nick_name: Option<String>,
    pub email: Option<String>,
    pub screen_name: Option<String>,
    pub app_installation_url: Option<String>,
    pub error_message: Option<String>,
    pub oauth_token_secret: Option<String>,
    pub need_confirmation: Option<bool>,
    pub oauth_id_token: Option<String>,
    pub full_name: Option<String>,
    pub display_name: Option<String>,
    pub kind: Option<String>,
    pub refresh_token: Option<String>,
    pub language: Option<String>,
    pub oauth_access_token: Option<String>,
    pub oauth_expire_in: Option<i32>,
    pub email_verified: Option<bool>,
    pub last_name: Option<String>,
    pub oauth_scope: Option<String>,
    pub oauth_request_token: Option<String>,
    pub oauth_authorization_code: Option<String>,
    pub original_email: Option<String>,
    pub provider_id: Option<String>,
    pub context: Option<String>,
    pub id_token: Option<String>,
    pub action: Option<String>,
    pub time_zone: Option<String>,
    pub email_recycled: Option<bool>,
}

Response of verifying the IDP assertion.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

verify assertion relyingparty (response)

Fields
federated_id: Option<String>

The unique ID identifies the IdP account.

local_id: Option<String>

The RP local ID if it's already been mapped to the IdP account identified by the federated ID.

photo_url: Option<String>

The URI of the public accessible profiel picture.

is_new_user: Option<bool>

True if it's a new user sign-in, false if it's a returning user.

input_email: Option<String>

It's the identifier param in the createAuthUri request if the identifier is an email. It can be used to check whether the user input email is different from the asserted email.

raw_user_info: Option<String>

Raw IDP-returned user info.

app_scheme: Option<String>

The custom scheme used by mobile app.

expires_in: Option<String>

If idToken is STS id token, then this field will be expiration time of STS id token in seconds.

verified_provider: Option<Vec<String>>

When action is 'map', contains the idps which can be used for confirmation.

need_email: Option<bool>

Whether need client to supply email to complete the federated login flow.

date_of_birth: Option<String>

The birth date of the IdP account.

first_name: Option<String>

The first name of the user.

nick_name: Option<String>

The nick name of the user.

email: Option<String>

The email returned by the IdP. NOTE: The federated login user may not own the email.

screen_name: Option<String>

The screen_name of a Twitter user or the login name at Github.

app_installation_url: Option<String>

URL for OTA app installation.

error_message: Option<String>

Client error code.

oauth_token_secret: Option<String>

The OAuth1 access token secret.

need_confirmation: Option<bool>

Whether the assertion is from a non-trusted IDP and need account linking confirmation.

oauth_id_token: Option<String>

The OIDC id token.

full_name: Option<String>

The full name of the user.

display_name: Option<String>

The display name of the user.

kind: Option<String>

The fixed string "identitytoolkit#VerifyAssertionResponse".

refresh_token: Option<String>

If idToken is STS id token, then this field will be refresh token.

language: Option<String>

The language preference of the user.

oauth_access_token: Option<String>

The OAuth2 access token.

oauth_expire_in: Option<i32>

The lifetime in seconds of the OAuth2 access token.

email_verified: Option<bool>

The value is true if the IDP is also the email provider. It means the user owns the email.

last_name: Option<String>

The last name of the user.

oauth_scope: Option<String>

The scope for the OpenID OAuth extension.

oauth_request_token: Option<String>

The user approved request token for the OpenID OAuth extension.

oauth_authorization_code: Option<String>

The OAuth2 authorization code.

original_email: Option<String>

The original email stored in the mapping storage. It's returned when the federated ID is associated to a different email.

provider_id: Option<String>

The IdP ID. For white listed IdPs it's a short domain name e.g. google.com, aol.com, live.net and yahoo.com. If the "providerId" param is set to OpenID OP identifer other than the whilte listed IdPs the OP identifier is returned. If the "identifier" param is federated ID in the createAuthUri request. The domain part of the federated ID is returned.

context: Option<String>

The opaque value used by the client to maintain context info between the authentication request and the IDP callback.

id_token: Option<String>

The ID token.

action: Option<String>

The action code.

time_zone: Option<String>

The timezone of the user.

email_recycled: Option<bool>

It's true if the email is recycled.